changes....
git-svn-id: https://keyvault-org.googlecode.com/svn/trunk@63 f5e68fdc-ee4c-11de-9b30-9f9d59cce923

+++ ssh.c
@@ -43,10 +43,10 @@ int tcp_connect(const char* hostname, int port) {
 	// Resolve the hostname
 	struct hostent* host = gethostbyname(hostname);
 	printf("\thostaddr: %u.%u.%u.%u (%u)\n", 
-		host->h_addr_list[0][0], 
-		host->h_addr_list[0][1], 
-		host->h_addr_list[0][2], 
-		host->h_addr_list[0][3], 
+		(uint8_t)host->h_addr_list[0][0], 
+		(uint8_t)host->h_addr_list[0][1], 
+		(uint8_t)host->h_addr_list[0][2], 
+		(uint8_t)host->h_addr_list[0][3], 
 		host->h_length);
 
 	// Create the TCP socket
